pyproject.toml
[project]
name = "nanabot"
version = "0.1.0"
description = "娜娜 Bot 是一个基于 NapCatQQ、NoneBot2 和 OneBot V11 的 QQ 机器人，允许 QQ 用户在私聊、群聊中与其互动"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "nonebot2[fastapi]>=2.4.3,<3.0.0",
    "nonebot-adapter-onebot>=2.4.6,<3.0.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
